+++ docs/changelog.md
@@ -31,6 +31,24 @@
 
 ## 更新说明
 
+1.7.4 fix: fix stock_analyst_rank_em interface
+
+    1. 修复 stock_analyst_rank_em 接口，规范输出格式
+
+1.7.3 fix: fix stock_profit_sheet_by_report_em interface
+
+    1. 修复 stock_profit_sheet_by_report_em 接口
+
+1.7.2 fix: fix futures_egg_price_yearly interface
+
+    1. 修复 futures_egg_price_yearly 接口
+    2. 修复 futures_egg_price 接口
+    3. 修复 futures_egg_price_area 接口
+
+1.7.1 fix: fix fund_etf_fund_info_em interface
+
+    1. 修复 fund_etf_fund_info_em 接口，提高访问速度并规范输出格式
+
 1.6.99 fix: fix macro_china_fx_gold interface
 
     1. 修复 macro_china_fx_gold 接口
@@ -1105,6 +1123,14 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.7.4 fix: fix stock_analyst_rank_em interface
+
+1.7.3 fix: fix stock_profit_sheet_by_report_em interface
+
+1.7.2 fix: fix futures_egg_price_yearly interface
+
+1.7.1 fix: fix fund_etf_fund_info_em interface
 
 1.6.99 fix: fix macro_china_fx_gold interface